Remove bayreuth and move IPv4 range to franken
"Freifunk Bayreuth" as a separate community seems to be dead for
several years. The last change on wiki.freifunk.net [1] happened
in 2012, and already in 2015 the bgp entry was removed from this
repo as it was dead [2].

No route is available for the IPv4 range, and nobody answers on
the e-mail address from [1].

Let's finally drop the entry for Bayreuth and move the IPv4 range
over to Franken, as that's the active community in that area.

[1] https://wiki.freifunk.net/index.php?title=Freifunk_Bayreuth&action=history
[2] 8bff35fe814d ("bayreuth: remove bgp block, has no active vpn")

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>